Se hace el mapeo d ebandas

+++ src/main/scala/LshMain.scala
@@ -3,6 +3,7 @@ import org.apache.spark.SparkContext
 
 import java.util.Calendar
 import java.io.File
+import java.util.function.ToIntFunction
 
 object LshMain {
   
@@ -30,7 +31,7 @@ object LshMain {
       //key = hash,group,signature, h0, D1, 2
       for(i <- 0 until 20) {   
         for(j <- 0 until 10) { 
-          array(i*10+j) = ("h"+i+",D"+j -> (i % 5).toString() )
+          array(i*10+j) = ((i+1).toString()+","+(j+1).toString() -> (i % 5).toString() )
         }
       }
 
@@ -46,14 +47,35 @@ object LshMain {
       
       if(!hdfs){
         val file = new File(directoryOutput)
-        file.listFiles().foreach { x => x.delete() }
+        if(file.listFiles() != null) file.listFiles().foreach { x => x.delete() }
         file.delete()       
       }else{
         val hadoopConf = new org.apache.hadoop.conf.Configuration()
         val hdfs = org.apache.hadoop.fs.FileSystem.get(new java.net.URI("hdfs://localhost:8020"), hadoopConf)
         hdfs.delete(new org.apache.hadoop.fs.Path(directoryOutput), true)        
       }     
-    }    
+    }  
+    
+    /**
+     * Mapea "id-shingle,id-doc,signature" to "band, id-doc", "id-shingle,id-doc,signature" Entera de Numero de linea/(Lineas por Banda)
+     */
+    def getBandAndDocument(line: String, linesPerBand: Long) : (String, String) = {
+      val line2 = line.replace("(", "").replace(")", "")
+      val fields = line2.split(",")
+      val nline = fields(0).toLong
+      (((nline / linesPerBand)+1).toString()+","+fields(1), line2)    
+    }
+    
+    /**
+     * Obtiene el bucket y con eso mapea a banda,bucket
+     */
+    def getBandAndBucket(key: String, value: Iterable[String]) : Array[(String, String)] = {
+      val array = new Array[(String, String)](5)
+      
+      value.foreach { x => ??? }
+      
+      array
+    }
     
     def main(args: Array[String]): Unit = {    
       val conf = new SparkConf().setAppName("matrix-mult2").setMaster("local")
@@ -66,8 +88,11 @@ object LshMain {
 //      val mat1FilePath = "hdfs://localhost:8020/user/matrix/matriz_a.dat"
 //      val mat2FilePath = "hdfs://localhost:8020/user/matrix/matriz_b-128.dat"
 //      val directoryOutput = "hdfs://localhost:8020/user/matrix/out"
+      
+      val hdfsFiles = false
       val signaturesFilePath = "data/signaturesMat"
       val directoryOutput = "data/out"
+      val rowsPerBand = 5
 //    val mat1FilePath = "/user/jonathan.medina/matrixMultiplication/matriz_a.dat"
 //    val mat2FilePath = "/user/jonathan.medina/matrixMultiplication/matriz_b-128.dat"
 //    val directoryOutput = "/user/jonathan.medina/out5"
@@ -76,16 +101,25 @@ object LshMain {
 //    val directoryOutput = "/user/jonathan.medina/out3"  
       /* *********************************************************************************************/
       
+     deleteDirectoryOutput(directoryOutput, hdfsFiles)
+//     createDataFile(sc, directoryOutput) 
 
-     //createDataFile(sc, directoryOutput) 
-     deleteDirectoryOutput(directoryOutput, false)
+    
      printlnWithTime("Inicia")
      
+     //lee archivo con tres valores: Id shingle(Long), Id documento (Long), Signature
      val file = sc.textFile(signaturesFilePath)
-     val map1 = file.zipWithIndex()
-     if(isDebug) map1.foreach(println)
+     if(isDebug) file.collect().foreach(line => println("linea: "+line))
      
-     map1.saveAsTextFile(directoryOutput)
+     /* Mapea banda Y documento
+      * "(1,1,0)", "(1,2,0)" map to: ("1,1", "1,1,0"),("1,2", "1,2,0")
+      */
+     val bandAndDocument = file.map(line => getBandAndDocument(line, rowsPerBand)).groupByKey()
+     if(isDebug) bandAndDocument.foreach(println)
+     
+     val bandAndBucket = bandAndDocument.flatMap{ case (k, v) => getBandAndBucket(k, v) }
+     if(isDebug) bandAndBucket.foreach(println)
+   
 
      printlnWithTime("Fin")
